+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - [BREAKING] Make Falcon DSA deterministic (#436).
 - `Smt.insert_inner_nodes` does not store empty subtrees ([#452](https://github.com/0xMiden/crypto/pull/452)).
 - Optimized `Smt::num_entries()` ([#455](https://github.com/0xMiden/crypto/pull/455)).
+- Disallow leaves with more than 2^16 entries ([#455](https://github.com/0xMiden/crypto/pull/455)).
 
 ## 0.15.6 (2025-07-15)